:bug: findByEmail 메소드 수정 및 input 제거
