--- src/main/java/com/project/routorybackend/account/dto/SignInInput.java
@@ -1,9 +1,0 @@
-package com.project.routorybackend.account.dto;
-
-import lombok.Data;
-
-@Data
-public class SignInInput {
-    private String email;
-    private String password;
-}
