--- src/main/java/com/project/routorybackend/account/dto/SignUpInput.java
@@ -1,22 +1,0 @@
-package com.project.routorybackend.account.dto;
-
-import com.project.routorybackend.account.model.Account;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class SignUpInput {
-    private String email;
-    private String password;
-    private String name;
-
-    public Account toAccount() {
-        return Account.builder()
-                .email(email)
-                .password(password)
-                .name(name)
-                .createdDate(LocalDateTime.now())
-                .build();
-    }
-}
