+++ src/main/java/com/project/routorybackend/account/dto/SignUpRequest.java
@@ -1,10 +1,13 @@
 package com.project.routorybackend.account.dto;
 
+import com.project.routorybackend.account.model.Account;
 import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
+import lombok.Data;
+
+import java.time.LocalDateTime;
 
 
-@Getter
+@Data
 public class SignUpRequest {
     @NotBlank(message = "빈문자열, 공백만 입력할 수 없습니다.")
     private String email;
@@ -13,12 +16,12 @@ public class SignUpRequest {
     @NotBlank(message = "빈문자열, 공백만 입력할 수 없습니다.")
     private String name;
 
-    public SignUpInput toInput() {
-        SignUpInput signUpInput = new SignUpInput();
-        signUpInput.setEmail(this.email);
-        signUpInput.setPassword(this.password);
-        signUpInput.setName(this.name);
-
-        return signUpInput;
+    public Account toAccount() {
+        return Account.builder()
+                .email(email)
+                .password(password)
+                .name(name)
+                .createdDate(LocalDateTime.now())
+                .build();
     }
 }
